refactor: split mcp-server.ts into focused modules under src/mcp/
Decompose the 1228-line mcp-server.ts into 9 focused modules:
- mcp/client.ts (296 lines) — AegisClient REST client + response types
- mcp/auth.ts (100 lines) — RBAC withAuth wrapper + role maps
- mcp/resources.ts (113 lines) — 4 MCP resource handlers
- mcp/tools/session-tools.ts (296 lines) — 12 session lifecycle tools
- mcp/tools/monitoring-tools.ts (142 lines) — 6 observability tools
- mcp/tools/pipeline-tools.ts (86 lines) — 3 batch/pipeline tools
- mcp/tools/management-tools.ts (81 lines) — 3 state management tools
- mcp/prompts.ts (141 lines) — 3 MCP prompt templates
- mcp/server.ts (50 lines) — createMcpServer orchestrator + stdio entry

mcp-server.ts becomes a 10-line re-export facade for backward compatibility.
All existing consumers (cli.ts, mcp-server.test.ts) continue to work unchanged.

Closes #1700
