+++ src/mcp-server.ts
@@ -1,1228 +1,10 @@
 /**
- * mcp-server.ts — MCP server mode for Aegis.
+ * mcp-server.ts — Re-export facade for backward compatibility.
  *
- * Exposes Aegis session orchestration as MCP tools via stdio transport.
- * CC sessions can natively discover and communicate with sibling sessions.
- *
- * Usage:
- *   aegis mcp                    # default port 9100
- *   aegis mcp --port 3000        # custom port
- *   claude mcp add --scope user aegis -- npx @onestepat4time/aegis mcp
- *
- * Issue #48: https://github.com/OneStepAt4time/aegis/issues/48
+ * The MCP server implementation lives in src/mcp/ modules.
+ * This file re-exports the public API so existing consumers
+ * (cli.ts, tests) continue to work without import changes.
  */
 
-import { McpServer, ResourceTemplate } from '@modelcontextprotocol/sdk/server/mcp.js';
-import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js';
-import type { ReadResourceResult } from '@modelcontextprotocol/sdk/types.js';
-import { z } from 'zod';
-import { readFileSync } from 'node:fs';
-import { dirname, join, resolve } from 'node:path';
-import { fileURLToPath } from 'node:url';
-
-import { isValidUUID } from './validation.js';
-import { type SessionInfo } from './session.js';
-import { type SessionMetrics, type SessionLatency, type SessionLatencySummary } from './metrics.js';
-import { type PipelineState, type BatchResult } from './pipeline.js';
-
-// Read version from package.json at startup (matches cli.ts pattern)
-const __dirname = dirname(fileURLToPath(import.meta.url));
-const pkg = JSON.parse(readFileSync(join(__dirname, '../package.json'), 'utf-8')) as { version: string };
-const VERSION: string = pkg.version;
-
-function normalizeWorkDirForCompare(workDir: string): string {
-  const isWindowsLikePath = /^[a-zA-Z]:[\\/]/.test(workDir) || workDir.startsWith('\\\\');
-  const normalizedPath = (isWindowsLikePath ? workDir : resolve(workDir))
-    .replace(/\\/g, '/')
-    .replace(/\/+$/, '');
-  return process.platform === 'win32' || isWindowsLikePath
-    ? normalizedPath.toLowerCase()
-    : normalizedPath;
-}
-
-function isSameOrChildWorkDir(candidate: string, parent: string): boolean {
-  const normalizedCandidate = normalizeWorkDirForCompare(candidate);
-  const normalizedParent = normalizeWorkDirForCompare(parent);
-  return normalizedCandidate === normalizedParent || normalizedCandidate.startsWith(`${normalizedParent}/`);
-}
-
-// ── API response types (Issue #577) ─────────────────────────────────
-
-interface ServerHealthResponse {
-  status: string;
-  version: string;
-  platform: NodeJS.Platform;
-  uptime: number;
-  sessions: { active: number; total: number };
-  tmux: { healthy: boolean; [key: string]: unknown };
-  timestamp: string;
-}
-
-interface CreateSessionResponse {
-  id: string;
-  windowName: string;
-  workDir: string;
-  status: string;
-  promptDelivery?: { delivered: boolean; attempts: number };
-  reused?: boolean;
-  [key: string]: unknown;
-}
-
-interface SendMessageResponse {
-  ok: boolean;
-  delivered: boolean;
-  attempts: number;
-  stall?: { stalled: true; types: string[] } | { stalled: false };
-}
-
-interface OkResponse {
-  ok: boolean;
-}
-
-interface CapturePaneResponse {
-  pane: string;
-}
-
-interface SessionLatencyResponse {
-  sessionId: string;
-  realtime: SessionLatency | null;
-  aggregated: SessionLatencySummary | null;
-}
-
-interface MemoryEntryResponse {
-  entry: {
-    key: string;
-    value: string;
-    namespace: string;
-    created_at: number;
-    updated_at: number;
-    expires_at?: number;
-  };
-}
-
-// ── Aegis REST client ───────────────────────────────────────────────
-
-export class AegisClient {
-  /** Cached role resolved from /v1/auth/verify. undefined = not yet resolved. */
-  private resolvedRole: string | undefined;
-
-  constructor(private baseUrl: string, private authToken?: string) {}
-
-  private validateSessionId(id: string): void {
-    if (!isValidUUID(id)) {
-      throw new Error(`Invalid session ID: ${id}`);
-    }
-  }
-
-  /**
-   * Resolve the RBAC role for the configured auth token.
-   * Calls POST /v1/auth/verify and caches only successful role resolutions.
-    * Returns 'admin' when no auth token is configured (matching server.ts behavior).
-    * Returns 'viewer' if role resolution fails or returns invalid data.
-   */
-  async resolveRole(): Promise<string> {
-    if (this.resolvedRole !== undefined) return this.resolvedRole;
-
-    if (!this.authToken) {
-      this.resolvedRole = 'admin';
-      return this.resolvedRole;
-    }
-
-    try {
-      const result = await this.request<{ valid: boolean; role?: string }>('/v1/auth/verify', {
-        method: 'POST',
-        body: JSON.stringify({ token: this.authToken }),
-      });
-      if (result.valid && (result.role === 'admin' || result.role === 'operator' || result.role === 'viewer')) {
-        this.resolvedRole = result.role;
-        return this.resolvedRole;
-      }
-      return 'viewer';
-    } catch {
-      return 'viewer';
-    }
-  }
-
-  private async request<T = unknown>(path: string, opts?: RequestInit): Promise<T> {
-    const hasBody = opts?.body !== undefined;
-    const headers: Record<string, string> = {
-      ...(hasBody ? { 'Content-Type': 'application/json' } : {}),
-      ...(this.authToken ? { Authorization: `Bearer ${this.authToken}` } : {}),
-    };
-    let res: Response;
-    try {
-      res = await fetch(`${this.baseUrl}${path}`, { ...opts, headers: { ...headers, ...opts?.headers } });
-    } catch (e: unknown) {
-      const cause = (e as { cause?: { code?: string } }).cause;
-      if (cause?.code === 'ECONNREFUSED') {
-        throw new Error('Aegis server is not running or not reachable');
-      }
-      throw new Error(`Network error: ${e instanceof Error ? e.message : String(e)}`);
-    }
-    if (!res.ok) {
-      const body = await res.json().catch(() => ({ error: res.statusText })) as { error?: string };
-      throw new Error(body.error || `HTTP ${res.status}`);
-    }
-    return res.json() as Promise<T>;
-  }
-
-  async listSessions(filter?: { status?: string; workDir?: string }): Promise<SessionInfo[]> {
-    const response = await this.request<{ sessions: SessionInfo[]; total: number }>('/v1/sessions');
-    let sessions = response.sessions;
-    if (filter?.status) {
-      sessions = sessions.filter((s) => s.status === filter.status);
-    }
-    if (filter?.workDir) {
-      sessions = sessions.filter((s) => isSameOrChildWorkDir(s.workDir, filter.workDir!));
-    }
-    return sessions;
-  }
-
-  async getSession(id: string): Promise<Record<string, unknown>> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}`);
-  }
-
-  async getHealth(id: string): Promise<Record<string, unknown>> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}/health`);
-  }
-
-  async getTranscript(id: string): Promise<Record<string, unknown>> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}/read`);
-  }
-
-  async sendMessage(id: string, text: string): Promise<SendMessageResponse> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}/send`, {
-      method: 'POST',
-      body: JSON.stringify({ text }),
-    });
-  }
-
-  async createSession(opts: { workDir: string; name?: string; prompt?: string }): Promise<CreateSessionResponse> {
-    return this.request('/v1/sessions', {
-      method: 'POST',
-      body: JSON.stringify(opts),
-    });
-  }
-
-  async killSession(id: string): Promise<OkResponse> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}`, {
-      method: 'DELETE',
-    });
-  }
-
-  async approvePermission(id: string): Promise<OkResponse> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}/approve`, {
-      method: 'POST',
-    });
-  }
-
-  async rejectPermission(id: string): Promise<OkResponse> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}/reject`, {
-      method: 'POST',
-    });
-  }
-
-  async getServerHealth(): Promise<ServerHealthResponse> {
-    return this.request('/v1/health');
-  }
-
-  async escapeSession(id: string): Promise<OkResponse> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}/escape`, {
-      method: 'POST',
-    });
-  }
-
-  async interruptSession(id: string): Promise<OkResponse> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}/interrupt`, {
-      method: 'POST',
-    });
-  }
-
-  async capturePane(id: string): Promise<CapturePaneResponse> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}/pane`);
-  }
-
-  async getSessionMetrics(id: string): Promise<SessionMetrics> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}/metrics`);
-  }
-
-  async getSessionSummary(id: string): Promise<Record<string, unknown>> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}/summary`);
-  }
-
-  async sendBash(id: string, command: string): Promise<OkResponse> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}/bash`, {
-      method: 'POST',
-      body: JSON.stringify({ command }),
-    });
-  }
-
-  async sendCommand(id: string, command: string): Promise<OkResponse> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}/command`, {
-      method: 'POST',
-      body: JSON.stringify({ command }),
-    });
-  }
-
-  async getSessionLatency(id: string): Promise<SessionLatencyResponse> {
-    this.validateSessionId(id);
-    return this.request(`/v1/sessions/${encodeURIComponent(id)}/latency`);
-  }
-
-  async batchCreateSessions(sessions: Array<{ workDir: string; name?: string; prompt?: string }>): Promise<BatchResult> {
-    return this.request('/v1/sessions/batch', {
-      method: 'POST',
-      body: JSON.stringify({ sessions }),
-    });
-  }
-
-  async listPipelines(): Promise<PipelineState[]> {
-    return this.request('/v1/pipelines');
-  }
-
-  async createPipeline(config: { name: string; workDir: string; steps: Array<{ name?: string; prompt: string }> }): Promise<PipelineState> {
-    return this.request('/v1/pipelines', {
-      method: 'POST',
-      body: JSON.stringify(config),
-    });
-  }
-
-  async getSwarm(): Promise<Record<string, unknown>> {
-    return this.request('/v1/swarm');
-  }
-
-  async setMemory(key: string, value: string, ttlSeconds?: number): Promise<MemoryEntryResponse> {
-    return this.request('/v1/memory', {
-      method: 'POST',
-      body: JSON.stringify({ key, value, ttlSeconds }),
-    });
-  }
-
-  async getMemory(key: string): Promise<MemoryEntryResponse> {
-    return this.request(`/v1/memory/${encodeURIComponent(key)}`);
-  }
-
-  async deleteMemory(key: string): Promise<OkResponse> {
-    return this.request(`/v1/memory/${encodeURIComponent(key)}`, {
-      method: 'DELETE',
-    });
-  }
-}
-
-// ── Error handling ──────────────────────────────────────────────────
-
-interface McpErrorEnvelope {
-  code: string;
-  message: string;
-  details?: unknown;
-}
-
-function formatToolError(e: unknown): { content: Array<{ type: 'text'; text: string }>; isError: true } {
-  if (e instanceof Error) {
-    let code: string;
-    if (e.message.includes('not running') || e.message.includes('not reachable') || e.message.includes('Network error')) {
-      code = 'SERVER_UNREACHABLE';
-    } else if (e.message.startsWith('Invalid session ID')) {
-      code = 'INVALID_SESSION_ID';
-    } else {
-      code = 'REQUEST_FAILED';
-    }
-    return {
-      content: [{ type: 'text' as const, text: JSON.stringify({ code, message: e.message } satisfies McpErrorEnvelope) }],
-      isError: true,
-    };
-  }
-  return {
-    content: [{ type: 'text' as const, text: JSON.stringify({ code: 'UNKNOWN_ERROR', message: String(e) } satisfies McpErrorEnvelope) }],
-    isError: true,
-  };
-}
-
-// ── MCP Tool Authorization (Issue #1407) ──────────────────────────────
-
-/** Minimum RBAC role required to call each MCP tool. */
-const TOOL_REQUIRED_ROLE: Record<string, string> = {
-  // viewer — read-only, no side effects
-  list_sessions: 'viewer',
-  get_status: 'viewer',
-  get_transcript: 'viewer',
-  server_health: 'viewer',
-  capture_pane: 'viewer',
-  get_session_metrics: 'viewer',
-  get_session_summary: 'viewer',
-  get_session_latency: 'viewer',
-  list_pipelines: 'viewer',
-  get_swarm: 'viewer',
-  state_get: 'viewer',
-  // operator — interactive but non-destructive
-  send_message: 'operator',
-  create_session: 'operator',
-  approve_permission: 'operator',
-  reject_permission: 'operator',
-  escape_session: 'operator',
-  interrupt_session: 'operator',
-  send_command: 'operator',
-  batch_create_sessions: 'operator',
-  create_pipeline: 'operator',
-  state_set: 'operator',
-  state_delete: 'operator',
-  // admin — destructive, requires elevated access
-  kill_session: 'admin',
-  send_bash: 'admin',
-};
-
-/** Numeric role levels for comparison. */
-const ROLE_LEVEL: Record<string, number> = {
-  admin: 3,
-  operator: 2,
-  viewer: 1,
-};
-
-function formatAuthError(toolName: string, role: string, required: string): { content: Array<{ type: 'text'; text: string }>; isError: true } {
-  return {
-    content: [{ type: 'text' as const, text: JSON.stringify({ code: 'FORBIDDEN', message: `Tool '${toolName}' requires '${required}' role, but token has '${role}' role` } satisfies McpErrorEnvelope) }],
-    isError: true,
-  };
-}
-
-/** Wrap a tool handler with per-tool role authorization. */
-function withAuth<TArgs>(
-  toolName: string,
-  handler: (args: TArgs) => Promise<{ content: Array<{ type: 'text'; text: string }>; isError?: boolean }>,
-  client: AegisClient,
-): (args: TArgs) => Promise<{ content: Array<{ type: 'text'; text: string }>; isError?: boolean }> {
-  return async (args) => {
-    const role = await client.resolveRole();
-    const required = TOOL_REQUIRED_ROLE[toolName];
-    if (required && (ROLE_LEVEL[role] ?? 0) < (ROLE_LEVEL[required] ?? 0)) {
-      return formatAuthError(toolName, role, required);
-    }
-    return handler(args);
-  };
-}
-
-// ── MCP Server ──────────────────────────────────────────────────────
-
-export function createMcpServer(aegisPort: number, authToken?: string): McpServer {
-  const client = new AegisClient(`http://127.0.0.1:${aegisPort}`, authToken);
-
-  const server = new McpServer(
-    { name: 'aegis', version: VERSION },
-    { capabilities: { tools: {}, resources: {} } },
-  );
-
-  // ── MCP Resources (Issue #442) ────────────────────────────────────
-
-  // aegis://sessions — compact session list
-  server.resource(
-    'sessions',
-    'aegis://sessions',
-    { description: 'List of active Aegis sessions (compact: id, name, status, workDir)', mimeType: 'application/json' },
-    async (uri): Promise<ReadResourceResult> => {
-      try {
-        const sessions = await client.listSessions();
-        const compact = sessions.map((s) => ({
-          id: s.id,
-          name: s.windowName,
-          status: s.status,
-          workDir: s.workDir,
-        }));
-        return {
-          contents: [{ uri: uri.href, mimeType: 'application/json', text: JSON.stringify(compact, null, 2) }],
-        };
-      } catch (e: unknown) {
-        const msg = e instanceof Error ? e.message : String(e);
-        return {
-          contents: [{ uri: uri.href, mimeType: 'text/plain', text: `Error: ${msg}` }],
-        };
-      }
-    },
-  );
-
-  // aegis://sessions/{id}/transcript — full JSONL transcript
-  server.resource(
-    'session-transcript',
-    new ResourceTemplate('aegis://sessions/{id}/transcript', { list: undefined }),
-    { description: 'Full JSONL transcript of an Aegis session', mimeType: 'application/json' },
-    async (uri, variables): Promise<ReadResourceResult> => {
-      const id = variables.id as string;
-      if (!isValidUUID(id)) {
-        return {
-          contents: [{ uri: uri.href, mimeType: 'text/plain', text: `Error: Invalid session ID: ${id}` }],
-        };
-      }
-      try {
-        const transcript = await client.getTranscript(id);
-        return {
-          contents: [{ uri: uri.href, mimeType: 'application/json', text: JSON.stringify(transcript, null, 2) }],
-        };
-      } catch (e: unknown) {
-        const msg = e instanceof Error ? e.message : String(e);
-        return {
-          contents: [{ uri: uri.href, mimeType: 'text/plain', text: `Error: ${msg}` }],
-        };
-      }
-    },
-  );
-
-  // aegis://sessions/{id}/pane — current terminal pane content
-  server.resource(
-    'session-pane',
-    new ResourceTemplate('aegis://sessions/{id}/pane', { list: undefined }),
-    { description: 'Current terminal pane content of an Aegis session', mimeType: 'text/plain' },
-    async (uri, variables): Promise<ReadResourceResult> => {
-      const id = variables.id as string;
-      if (!isValidUUID(id)) {
-        return {
-          contents: [{ uri: uri.href, mimeType: 'text/plain', text: `Error: Invalid session ID: ${id}` }],
-        };
-      }
-      try {
-        const result = await client.capturePane(id);
-        const text = typeof result.pane === 'string' ? result.pane : JSON.stringify(result, null, 2);
-        return {
-          contents: [{ uri: uri.href, mimeType: 'text/plain', text }],
-        };
-      } catch (e: unknown) {
-        const msg = e instanceof Error ? e.message : String(e);
-        return {
-          contents: [{ uri: uri.href, mimeType: 'text/plain', text: `Error: ${msg}` }],
-        };
-      }
-    },
-  );
-
-  // aegis://health — server health status
-  server.resource(
-    'health',
-    'aegis://health',
-    { description: 'Aegis server health status (version, uptime, session counts)', mimeType: 'application/json' },
-    async (uri): Promise<ReadResourceResult> => {
-      try {
-        const health = await client.getServerHealth();
-        return {
-          contents: [{ uri: uri.href, mimeType: 'application/json', text: JSON.stringify(health, null, 2) }],
-        };
-      } catch (e: unknown) {
-        const msg = e instanceof Error ? e.message : String(e);
-        return {
-          contents: [{ uri: uri.href, mimeType: 'text/plain', text: `Error: ${msg}` }],
-        };
-      }
-    },
-  );
-
-  // ── list_sessions ──
-  server.tool(
-    'list_sessions',
-    'List Aegis-managed Claude Code sessions. Optionally filter by status or workDir substring.',
-    {
-      status: z.string().optional().describe('Filter by status (e.g., idle, working, permission_prompt)'),
-      workDir: z.string().optional().describe('Filter by workDir substring (e.g., "my-project")'),
-    },
-    withAuth('list_sessions', async ({ status, workDir }) => {
-      try {
-        const sessions = await client.listSessions({ status, workDir });
-        const summary = sessions.map((s) => ({
-          id: s.id,
-          name: s.windowName,
-          status: s.status,
-          workDir: s.workDir,
-          createdAt: new Date(s.createdAt).toISOString(),
-          lastActivity: new Date(s.lastActivity).toISOString(),
-        }));
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(summary, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── get_status ──
-  server.tool(
-    'get_status',
-    'Get detailed status and health of a specific Aegis session.',
-    {
-      sessionId: z.string().describe('The session ID to check'),
-    },
-    withAuth('get_status', async ({ sessionId }) => {
-      try {
-        const [session, health] = await Promise.all([
-          client.getSession(sessionId),
-          client.getHealth(sessionId).catch(() => null),
-        ]);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify({ ...session, health }, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── get_transcript ──
-  server.tool(
-    'get_transcript',
-    'Read the conversation transcript of another Aegis session. Returns recent messages from the JSONL log.',
-    {
-      sessionId: z.string().describe('The session ID to read from'),
-    },
-    withAuth('get_transcript', async ({ sessionId }) => {
-      try {
-        const transcript = await client.getTranscript(sessionId);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(transcript, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── send_message ──
-  server.tool(
-    'send_message',
-    'Send a message to another Aegis session. The message is delivered via tmux send-keys with delivery verification. Returns stall information if the session is currently stalled.',
-    {
-      sessionId: z.string().describe('The target session ID'),
-      text: z.string().describe('The message text to send'),
-    },
-    withAuth('send_message', async ({ sessionId, text }) => {
-      try {
-        const result = await client.sendMessage(sessionId, text);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── create_session ──
-  server.tool(
-    'create_session',
-    'Spawn a new Claude Code session managed by Aegis. Returns the session ID and initial status.',
-    {
-      workDir: z.string().describe('Working directory for the new session'),
-      name: z.string().optional().describe('Optional human-readable name for the session'),
-      prompt: z.string().optional().describe('Optional initial prompt to send after creation'),
-    },
-    withAuth('create_session', async ({ workDir, name, prompt }) => {
-      try {
-        const session = await client.createSession({ workDir, name, prompt });
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify({
-              id: session.id,
-              name: session.windowName,
-              status: 'created',
-              workDir: session.workDir,
-              promptDelivery: session.promptDelivery,
-            }, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── kill_session ──
-  server.tool(
-    'kill_session',
-    'Kill an Aegis session. Deletes the tmux window and cleans up all resources.',
-    {
-      sessionId: z.string().describe('The session ID to kill'),
-    },
-    withAuth('kill_session', async ({ sessionId }) => {
-      try {
-        const result = await client.killSession(sessionId);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── approve_permission ──
-  server.tool(
-    'approve_permission',
-    'Approve a pending permission prompt in an Aegis session.',
-    {
-      sessionId: z.string().describe('The session ID with a pending permission prompt'),
-    },
-    withAuth('approve_permission', async ({ sessionId }) => {
-      try {
-        const result = await client.approvePermission(sessionId);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── reject_permission ──
-  server.tool(
-    'reject_permission',
-    'Reject a pending permission prompt in an Aegis session.',
-    {
-      sessionId: z.string().describe('The session ID with a pending permission prompt'),
-    },
-    withAuth('reject_permission', async ({ sessionId }) => {
-      try {
-        const result = await client.rejectPermission(sessionId);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── server_health ──
-  server.tool(
-    'server_health',
-    'Check the health and status of the Aegis server. Returns version, uptime, and session counts.',
-    {},
-    withAuth('server_health', async () => {
-      try {
-        const result = await client.getServerHealth();
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── escape_session ──
-  server.tool(
-    'escape_session',
-    'Send an Escape keypress to an Aegis session. Useful for dismissing prompts or cancelling operations.',
-    {
-      sessionId: z.string().describe('The session ID to send escape to'),
-    },
-    withAuth('escape_session', async ({ sessionId }) => {
-      try {
-        const result = await client.escapeSession(sessionId);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── interrupt_session ──
-  server.tool(
-    'interrupt_session',
-    'Send Ctrl+C to interrupt the current operation in an Aegis session.',
-    {
-      sessionId: z.string().describe('The session ID to interrupt'),
-    },
-    withAuth('interrupt_session', async ({ sessionId }) => {
-      try {
-        const result = await client.interruptSession(sessionId);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── capture_pane ──
-  server.tool(
-    'capture_pane',
-    'Capture the raw terminal pane content of an Aegis session. Returns the current visible text.',
-    {
-      sessionId: z.string().describe('The session ID to capture'),
-    },
-    withAuth('capture_pane', async ({ sessionId }) => {
-      try {
-        const result = await client.capturePane(sessionId);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── get_session_metrics ──
-  server.tool(
-    'get_session_metrics',
-    'Get performance metrics for a specific Aegis session (message counts, latency, etc.).',
-    {
-      sessionId: z.string().describe('The session ID to get metrics for'),
-    },
-    withAuth('get_session_metrics', async ({ sessionId }) => {
-      try {
-        const result = await client.getSessionMetrics(sessionId);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── get_session_summary ──
-  server.tool(
-    'get_session_summary',
-    'Get a summary of an Aegis session including message counts, duration, and status history.',
-    {
-      sessionId: z.string().describe('The session ID to summarize'),
-    },
-    withAuth('get_session_summary', async ({ sessionId }) => {
-      try {
-        const result = await client.getSessionSummary(sessionId);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── send_bash ──
-  server.tool(
-    'send_bash',
-    'Execute a bash command in an Aegis session. The command is prefixed with "!" and sent via tmux.',
-    {
-      sessionId: z.string().describe('The session ID to send the bash command to'),
-      command: z.string().describe('The bash command to execute'),
-    },
-    withAuth('send_bash', async ({ sessionId, command }) => {
-      try {
-        const result = await client.sendBash(sessionId, command);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── send_command ──
-  server.tool(
-    'send_command',
-    'Send a slash command to an Aegis session. The command is prefixed with "/" if not already.',
-    {
-      sessionId: z.string().describe('The session ID to send the command to'),
-      command: z.string().describe('The slash command to send (e.g., "help", "compact")'),
-    },
-    withAuth('send_command', async ({ sessionId, command }) => {
-      try {
-        const result = await client.sendCommand(sessionId, command);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── get_session_latency ──
-  server.tool(
-    'get_session_latency',
-    'Get latency metrics for a specific Aegis session, including realtime and aggregated measurements.',
-    {
-      sessionId: z.string().describe('The session ID to get latency for'),
-    },
-    withAuth('get_session_latency', async ({ sessionId }) => {
-      try {
-        const result = await client.getSessionLatency(sessionId);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── batch_create_sessions ──
-  server.tool(
-    'batch_create_sessions',
-    'Create multiple Aegis sessions in a single batch operation.',
-    {
-      sessions: z.array(z.object({
-        workDir: z.string().describe('Working directory for the session'),
-        name: z.string().optional().describe('Optional human-readable name'),
-        prompt: z.string().optional().describe('Optional initial prompt'),
-      })).min(1).max(50).describe('Array of session specifications to create (max 50)'),
-    },
-    withAuth('batch_create_sessions', async ({ sessions: sessionSpecs }) => {
-      try {
-        const result = await client.batchCreateSessions(sessionSpecs);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── list_pipelines ──
-  server.tool(
-    'list_pipelines',
-    'List all configured pipelines in the Aegis server.',
-    {},
-    withAuth('list_pipelines', async () => {
-      try {
-        const result = await client.listPipelines();
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── create_pipeline ──
-  server.tool(
-    'create_pipeline',
-    'Create a new pipeline for orchestrating multiple Aegis sessions in sequence.',
-    {
-      name: z.string().describe('Name of the pipeline'),
-      workDir: z.string().describe('Working directory for pipeline sessions'),
-      steps: z.array(z.object({
-        name: z.string().optional().describe('Step name'),
-        prompt: z.string().describe('Prompt for this step'),
-      })).min(1).max(50).describe('Array of pipeline steps (max 50)'),
-    },
-    withAuth('create_pipeline', async ({ name, workDir, steps }) => {
-      try {
-        const result = await client.createPipeline({ name, workDir, steps });
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── get_swarm ──
-  server.tool(
-    'get_swarm',
-    'Get a snapshot of all Claude Code processes detected on the system (the "swarm").',
-    {},
-    withAuth('get_swarm', async () => {
-      try {
-        const result = await client.getSwarm();
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── state_set ──
-  server.tool(
-    'state_set',
-    'Set a shared state key/value entry via Aegis memory bridge.',
-    {
-      key: z.string().describe('State key in namespace/key format (e.g., pipeline/run-123)'),
-      value: z.string().describe('State payload as string'),
-      ttlSeconds: z.number().int().positive().max(86400 * 30).optional().describe('Optional TTL in seconds (max 30 days)'),
-    },
-    withAuth('state_set', async ({ key, value, ttlSeconds }) => {
-      try {
-        const result = await client.setMemory(key, value, ttlSeconds);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── state_get ──
-  server.tool(
-    'state_get',
-    'Get a shared state key/value entry via Aegis memory bridge.',
-    {
-      key: z.string().describe('State key in namespace/key format (e.g., pipeline/run-123)'),
-    },
-    withAuth('state_get', async ({ key }) => {
-      try {
-        const result = await client.getMemory(key);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── state_delete ──
-  server.tool(
-    'state_delete',
-    'Delete a shared state key/value entry via Aegis memory bridge.',
-    {
-      key: z.string().describe('State key in namespace/key format (e.g., pipeline/run-123)'),
-    },
-    withAuth('state_delete', async ({ key }) => {
-      try {
-        const result = await client.deleteMemory(key);
-        return {
-          content: [{
-            type: 'text' as const,
-            text: JSON.stringify(result, null, 2),
-          }],
-        };
-      } catch (e: unknown) {
-        return formatToolError(e);
-      }
-    }, client),
-  );
-
-  // ── MCP Prompts (Issue #443) ────────────────────────────────────────
-
-  server.prompt(
-    'implement_issue',
-    'Create a session and generate a structured implementation prompt for a GitHub issue.',
-    {
-      issueNumber: z.string().describe('GitHub issue number'),
-      workDir: z.string().describe('Working directory for the new session'),
-      repoOwner: z.string().optional().describe('Repository owner (e.g., "OneStepAt4time")'),
-      repoName: z.string().optional().describe('Repository name (e.g., "aegis")'),
-    },
-    async ({ issueNumber, workDir, repoOwner, repoName }) => {
-      const owner = repoOwner || 'OneStepAt4time';
-      const repo = repoName || 'aegis';
-      const issueUrl = `https://github.com/${owner}/${repo}/issues/${issueNumber}`;
-
-      return {
-        messages: [
-          {
-            role: 'user' as const,
-            content: {
-              type: 'text' as const,
-              text: [
-                'You are tasked with implementing a GitHub issue.',
-                '',
-                `Issue: ${owner}/${repo}#${issueNumber}`,
-                `URL: ${issueUrl}`,
-                `Working directory: ${workDir}`,
-                '',
-                'Steps:',
-                `1. Create a new Aegis session in ${workDir}`,
-                `2. Read the GitHub issue at ${issueUrl} to understand the requirements`,
-                '3. Analyze the codebase to understand the current architecture',
-                '4. Plan the implementation approach',
-                '5. Implement the changes following project conventions',
-                '6. Run the quality gate: npx tsc --noEmit && npm run build && npm test',
-                '7. If tests pass, commit with a conventional commit message',
-                '',
-                'Use the create_session tool to start, then send_message for each step.',
-              ].join('\n'),
-            },
-          },
-        ],
-      };
-    },
-  );
-
-  server.prompt(
-    'review_pr',
-    'Create a session and generate a structured code review prompt for a GitHub pull request.',
-    {
-      prNumber: z.string().describe('GitHub pull request number'),
-      workDir: z.string().describe('Working directory for the new session'),
-      repoOwner: z.string().optional().describe('Repository owner (e.g., "OneStepAt4time")'),
-      repoName: z.string().optional().describe('Repository name (e.g., "aegis")'),
-    },
-    async ({ prNumber, workDir, repoOwner, repoName }) => {
-      const owner = repoOwner || 'OneStepAt4time';
-      const repo = repoName || 'aegis';
-      const prUrl = `https://github.com/${owner}/${repo}/pull/${prNumber}`;
-
-      return {
-        messages: [
-          {
-            role: 'user' as const,
-            content: {
-              type: 'text' as const,
-              text: [
-                'You are tasked with reviewing a GitHub pull request.',
-                '',
-                `PR: ${owner}/${repo}#${prNumber}`,
-                `URL: ${prUrl}`,
-                `Working directory: ${workDir}`,
-                '',
-                'Steps:',
-                `1. Create a new Aegis session in ${workDir}`,
-                `2. Fetch the PR details: gh pr view ${prNumber} --repo ${owner}/${repo}`,
-                `3. Fetch the PR diff: gh pr diff ${prNumber} --repo ${owner}/${repo}`,
-                '4. Review the changes for:',
-                '   - Correctness and edge cases',
-                '   - Adherence to project coding conventions (see CLAUDE.md)',
-                '   - Security vulnerabilities (injection, XSS, etc.)',
-                '   - Test coverage for new code',
-                '   - Breaking changes or backwards compatibility',
-                '5. Post the review as a PR comment using gh api',
-                '',
-                'Use the create_session tool to start, then send_message for each step.',
-              ].join('\n'),
-            },
-          },
-        ],
-      };
-    },
-  );
-
-  server.prompt(
-    'debug_session',
-    'Generate a diagnostic summary for an Aegis session by reading its transcript and status.',
-    {
-      sessionId: z.string().describe('The Aegis session ID to debug'),
-    },
-    async ({ sessionId }) => {
-      return {
-        messages: [
-          {
-            role: 'user' as const,
-            content: {
-              type: 'text' as const,
-              text: [
-                'You are diagnosing an Aegis session that may be stuck or misbehaving.',
-                '',
-                `Session ID: ${sessionId}`,
-                '',
-                'Steps:',
-                `1. Get the session status using get_status for session ${sessionId}`,
-                `2. Read the transcript using get_transcript for session ${sessionId}`,
-                `3. Capture the current terminal pane using capture_pane for session ${sessionId}`,
-                '4. Analyze the findings:',
-                '   - Is the session in an unexpected state (permission_prompt, unknown)?',
-                '   - Are there error messages in the transcript?',
-                '   - Is the session stalled (no recent activity)?',
-                '   - Are there repeated permission requests?',
-                '5. Provide a diagnostic summary with recommended actions',
-                '',
-                'Use get_status, get_transcript, and capture_pane tools to gather data.',
-              ].join('\n'),
-            },
-          },
-        ],
-      };
-    },
-  );
-
-  return server;
-}
-
-// ── Main (stdio entrypoint) ─────────────────────────────────────────
-
-export async function startMcpServer(port: number, authToken?: string): Promise<void> {
-  const server = createMcpServer(port, authToken);
-  const transport = new StdioServerTransport();
-  await server.connect(transport);
-  // Server runs until stdin closes
-}
+export { AegisClient } from './mcp/client.js';
+export { createMcpServer, startMcpServer } from './mcp/server.js';
